+++ spec/rubocop/cop/gemspec/required_ruby_version_spec.rb
@@ -2,11 +2,19 @@
 
 RSpec.describe RuboCop::Cop::Gemspec::RequiredRubyVersion, :config do
   context 'target ruby version > 2.7', :ruby27 do
-    it 'registers an offense when `required_ruby_version` is lower than ' \
-       '`TargetRubyVersion`' do
+    it 'registers an offense when `required_ruby_version` is specified with >= and is lower than `TargetRubyVersion`' do
       expect_offense(<<~RUBY, '/path/to/foo.gemspec')
         Gem::Specification.new do |spec|
           spec.required_ruby_version = '>= 2.6.0'
+                                       ^^^^^^^^^^ `required_ruby_version` (2.6, declared in foo.gemspec) and `TargetRubyVersion` (2.7, which may be specified in .rubocop.yml) should be equal.
+        end
+      RUBY
+    end
+
+    it 'registers an offense when `required_ruby_version` is specified with ~> and is lower than `TargetRubyVersion`' do
+      expect_offense(<<~RUBY, '/path/to/foo.gemspec')
+        Gem::Specification.new do |spec|
+          spec.required_ruby_version = '~> 2.6.0'
                                        ^^^^^^^^^^ `required_ruby_version` (2.6, declared in foo.gemspec) and `TargetRubyVersion` (2.7, which may be specified in .rubocop.yml) should be equal.
         end
       RUBY
@@ -37,8 +45,7 @@ RSpec.describe RuboCop::Cop::Gemspec::RequiredRubyVersion, :config do
   end
 
   context 'target ruby version > 2.5', :ruby25 do
-    it 'registers an offense when `required_ruby_version` is higher than ' \
-       '`TargetRubyVersion`' do
+    it 'registers an offense when `required_ruby_version` is specified with >= and is higher than `TargetRubyVersion`' do
       expect_offense(<<~RUBY, '/path/to/bar.gemspec')
         Gem::Specification.new do |spec|
           spec.required_ruby_version = '>= 2.6.0'
@@ -46,11 +53,19 @@ RSpec.describe RuboCop::Cop::Gemspec::RequiredRubyVersion, :config do
         end
       RUBY
     end
+
+    it 'registers an offense when `required_ruby_version` is specified with ~> and is higher than `TargetRubyVersion`' do
+      expect_offense(<<~RUBY, '/path/to/bar.gemspec')
+        Gem::Specification.new do |spec|
+          spec.required_ruby_version = '~> 2.6.0'
+                                       ^^^^^^^^^^ `required_ruby_version` (2.6, declared in bar.gemspec) and `TargetRubyVersion` (2.5, which may be specified in .rubocop.yml) should be equal.
+        end
+      RUBY
+    end
   end
 
   context 'target ruby version > 2.6', :ruby26 do
-    it 'does not register an offense when `required_ruby_version` equals ' \
-       '`TargetRubyVersion`' do
+    it 'does not register an offense when `required_ruby_version` is specified with >= and equals `TargetRubyVersion`' do
       expect_no_offenses(<<~RUBY)
         Gem::Specification.new do |spec|
           spec.required_ruby_version = '>= 2.6.0'
@@ -58,11 +73,28 @@ RSpec.describe RuboCop::Cop::Gemspec::RequiredRubyVersion, :config do
       RUBY
     end
 
-    it 'does not register an offense when `required_ruby_version` ' \
-       '(omit patch version) equals `TargetRubyVersion`' do
+    it 'does not register an offense when `required_ruby_version` is specified with ~> and equals `TargetRubyVersion`' do
+      expect_no_offenses(<<~RUBY)
+        Gem::Specification.new do |spec|
+          spec.required_ruby_version = '~> 2.6.0'
+        end
+      RUBY
+    end
+
+    it 'does not register an offense when `required_ruby_version` is specified with >= without a patch version and ' \
+       'equals `TargetRubyVersion`' do
       expect_no_offenses(<<~RUBY)
         Gem::Specification.new do |spec|
           spec.required_ruby_version = '>= 2.6'
+        end
+      RUBY
+    end
+
+    it 'does not register an offense when `required_ruby_version` is specified with ~> without a patch version and ' \
+       'equals `TargetRubyVersion`' do
+      expect_no_offenses(<<~RUBY)
+        Gem::Specification.new do |spec|
+          spec.required_ruby_version = '~> 2.6'
         end
       RUBY
     end
@@ -72,6 +104,26 @@ RSpec.describe RuboCop::Cop::Gemspec::RequiredRubyVersion, :config do
       expect_no_offenses(<<~RUBY)
         Gem::Specification.new do |spec|
           spec.required_ruby_version = ['>= 2.6.0', '< 2.8.0']
+        end
+      RUBY
+    end
+
+    it 'registers an offense when `required_ruby_version` is specified with >= without a minor version and is lower ' \
+       'than `TargetRubyVersion`' do
+      expect_offense(<<~RUBY, 'bar.gemspec')
+        Gem::Specification.new do |spec|
+          spec.required_ruby_version = '>= 2'
+                                       ^^^^^^ `required_ruby_version` (2, declared in bar.gemspec) and `TargetRubyVersion` (2.6, which may be specified in .rubocop.yml) should be equal.
+        end
+      RUBY
+    end
+
+    it 'registers an offense when `required_ruby_version` is specified with ~> without a minor version and is lower ' \
+       'than `TargetRubyVersion`' do
+      expect_offense(<<~RUBY, 'bar.gemspec')
+        Gem::Specification.new do |spec|
+          spec.required_ruby_version = '~> 2'
+                                       ^^^^^^ `required_ruby_version` (2, declared in bar.gemspec) and `TargetRubyVersion` (2.6, which may be specified in .rubocop.yml) should be equal.
         end
       RUBY
     end